Implementando os metodos GET e POST

+++ src/main/java/com/produtos/apirest/models/Produto.java
@@ -16,17 +16,16 @@ public class Produto implements Serializable {
 	
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO) 
-	private Integer id;
+	private long id;
 
 	private String nome;
 	private BigDecimal quantidade;
 	private BigDecimal valor;
-	
-	
-	public Integer getId() {
+
+	public long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getNome() {
